Rough cut of using typed rpc in xi-core
This will hopefully illustrate the merits (and limitations) of #393.
There are a few warty bits that aren't very well explained, like the
`EmptyStruct` type (which can be removed with a minor protocol change),
and the single included test should be broken up and documentation
improved before this is considered ready.